Create ReverseString.java
This program reverses a string using a two pointer approach where one of the pointer named start is at the start of the string and the other one named end is at the end of the string. I swap the letters at start and end then increment start by one and then decrement end by one until start becomes greater than end(which means that the string has been reversed).