create LonelyInterger.java
in Lonely Integer program you have to find that Integer that is not repeating or that is unique in array. so for this you have to first  sort the array to find unique integer so i sort the array with existing function of Arrays in java then with logic if nteger present in array is equal to i+1 so it is not integer if not equal so it is integer. and return the Integer